BUG: fix Ambiguity in insert call

+++ Source/cmLocalVisualStudio7Generator.cxx
@@ -545,7 +545,9 @@ void cmLocalVisualStudio7Generator::FillFlagMapFromCommandFlags(
     {
     option.reserve(strlen(flagTable->commandFlag+2));
     // first do the - version
-    option.insert(0, 1, '-');
+    option.insert(static_cast<unsigned int>(0), 
+                  static_cast<unsigned int>(1), 
+                  '-');
     option.append(flagTable->commandFlag);
     while(flags.find(option) != flags.npos)
       {
